Deleted deprecated converters
